Make downloading scripts resumable.
At present if we loose network halfway through downloading a set of files, we
have to restart from the beginning, because there is no way to ensure that we
only download stuff not already downloaded. Adding the job_guid to the filename
allows us to skip files that are already downloaded, assuming we take care to
rename the file after the download is complete to avoid using partial logs.