Rewrite to use taskcluster directly not treeherder
This was prinipcally motivated by some treeherder HTTP responses that
were tripping up reqwest and resulting in no logs being
downloaded. But it also means that we rely on one fewer service and
don't depend on unstable API endpoints.

The main issue with the rewrite is that we require one request per job
to get the list of artifacts and another to get the actual artifact,
which slows things down a little compared to the treeherder-based
setup. We also have to consult hg to convert the revision from prefix
form to a full revision we can use to get the decision task.