Fetch all taskgroups from GitHub
When jobs are added to a GitHub push e.g. because the branch is
tagged, we end up running multiple `decision-task` tasks. Previusly we
were only fetching the latest of these, so we could only download
artifacts from the tasks associated with the latest jobs.

To fix this, we ask github to produce a list of all tasks for a
specific job, not just the latest ones. We also update the
get_taskgroup method to allow returning multiple taskgroups, and then
fetch the tasks from all the taskgroups.